+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/core/view/ui/fragment/FragmentExtensions.kt
@@ -1,5 +1,6 @@
 package org.hyperskill.app.android.core.view.ui.fragment
 
+import androidx.annotation.IdRes
 import androidx.fragment.app.Fragment
 
 @Suppress("UNCHECKED_CAST")
@@ -19,3 +20,20 @@ fun <T> Fragment.parentOfType(klass: Class<T>): T? =
                 null
             }
         }
+
+/**
+ * Set fragment built by [buildFragment] into [containerViewId] using Fragment.childFragmentManager
+ * */
+fun Fragment.setChildFragment(
+    @IdRes containerViewId: Int,
+    tag: String,
+    buildFragment: () -> Fragment
+) {
+    if (childFragmentManager.findFragmentByTag(tag) == null) {
+        val fragment = buildFragment()
+        childFragmentManager
+            .beginTransaction()
+            .add(containerViewId, fragment, tag)
+            .commitNow()
+    }
+}
