+++ shared/src/commonMain/kotlin/org/hyperskill/app/core/presentation/TransformState.kt
@@ -3,11 +3,11 @@ package org.hyperskill.app.core.presentation
 import ru.nobird.app.core.model.Cancellable
 import ru.nobird.app.presentation.redux.feature.Feature
 
-fun <State1, State2, Message, Action> Feature<State1, Message, Action>.transformState(
-    transformState: (State1) -> State2
-): Feature<State2, Message, Action>  =
-    object : Feature<State2, Message, Action> {
-        override val state: State2
+fun <InitialState, NewState, Message, Action> Feature<InitialState, Message, Action>.transformState(
+    transformState: (InitialState) -> NewState
+): Feature<NewState, Message, Action>  =
+    object : Feature<NewState, Message, Action> {
+        override val state: NewState
             get() = transformState(this@transformState.state)
 
         override fun cancel() {
@@ -17,7 +17,7 @@ fun <State1, State2, Message, Action> Feature<State1, Message, Action>.transform
         override fun addActionListener(listener: (action: Action) -> Unit): Cancellable =
             this@transformState.addActionListener(listener)
 
-        override fun addStateListener(listener: (state2: State2) -> Unit): Cancellable =
+        override fun addStateListener(listener: (state2: NewState) -> Unit): Cancellable =
             this@transformState.addStateListener { state1 ->
                 listener(transformState(state1))
             }
